+++ public/scripts/collisions.js
@@ -4,169 +4,183 @@
 class Collisions {
   // Constructor will take the player and baddies separately, compare their positions each cycle, and update the collision status of
   // any sprites involved in a collision (i.e. the player + one or more baddies):
-  constructor(player, baddies) {
-    this.baddies = baddies;
+  constructor(player) {
     this.player = player;
   }
 
   // DF method will determine who faces who in collisions that occur between player and baddie:
-  determineFacing(deltaX, player, baddie) {
+  determineFacing(deltaX, deltaEu, player, baddie) {
     let absDeltaX = Math.abs(deltaX);
-    console.log(
-      `player faces ${player.facing} at ${absDeltaX} range from baddie, who faces ${baddie.facing}`
-    );
     // Enter the logic tree: If delta x is greater than zero, player is to the right of the baddie:
     if (deltaX > 0) {
       if (player.facing === 'right') {
-        // If you're to the right, and you and the enemy face right, enemy is behind you:
         if (baddie.facing === 'right') {
-          console.log('they came from behind!');
-          // For every situation, check if player is attacking and if their distance is within the appropriate range:
+          // Player is to the right, player faces right, baddie faces right: baddie is behind you!
+          // If enemy is behind you then it doesn't matter if you're attacking:
           if (
-            player.isAttacking &&
-            absDeltaX >
-              badDictionary[baddie.id].baddieBehind.attackSuccessRange[0] &&
-            absDeltaX <
-              badDictionary[baddie.id].baddieBehind.attackSuccessRange[1]
+            deltaEu <
+            badDictionary[`baddie_${baddie.type}`].baddieBehind.killRadius
           ) {
-            // if they are, the baddie is killed!
-            baddie.collisionStatus = 'hitFromFunnyAngle';
-            console.log('kill successful at range', absDeltaX);
-          } else if (
-            absDeltaX < badDictionary[baddie.id].baddieBehind.killRadius
-          ) {
-            player.collisionStatus = 'impact';
+            player.collisionStatus = `impact-with-${baddie.id}`;
+            console.log('they came from... behind!');
+            console.log(
+              `your position: ${player.x}, ${player.y}; their position: ${baddie.x}, ${baddie.y}, euclid: ${deltaEu}`
+            );
           }
         } else {
-          console.log('back to back');
+          // Player is to the right, player faces right, baddie faces left: you stand back-to-back.
+          // if you and baddie stand back to back then it doesn't matter if you're attacking:
           if (
-            player.isAttacking &&
-            absDeltaX >
-              badDictionary[baddie.id].backToBack.attackSuccessRange[0] &&
-            absDeltaX <
-              badDictionary[baddie.id].backToBack.attackSuccessRange[1]
+            deltaEu <
+            badDictionary[`baddie_${baddie.type}`].backToBack.killRadius
           ) {
-            // if they are, the baddie is killed!
-            baddie.collisionStatus = 'hitFromElsewhere';
-            console.log('kill successful at range', absDeltaX);
-          } else if (
-            absDeltaX < badDictionary[baddie.id].backToBack.killRadius
-          ) {
-            player.collisionStatus = 'impact';
+            player.collisionStatus = `impact-with-${baddie.id}`;
+            console.log('back to back');
+            console.log(
+              `your position: ${player.x}, ${player.y}; their position: ${baddie.x}, ${baddie.y}, euclid: ${deltaEu}`
+            );
           }
         }
-        // player faces left:
+        // Player faces left:
       } else {
         if (baddie.facing === 'right') {
-          console.log('face to face');
+          // Player is to the right, player faces left, baddie faces right: you face each other:
+          // Since you face the baddie, attacking IS possible:
           if (
+            // Player must be in the attacking state to consider whether a strike is successful:
             player.isAttacking &&
-            absDeltaX >
-              badDictionary[baddie.id].faceToFace.attackSuccessRange[0] &&
-            absDeltaX <
-              badDictionary[baddie.id].faceToFace.attackSuccessRange[1]
+            // Distance-to-strike calculus: Strike succeeds if (player attack radius > [Euclidean distance MINUS baddie WIDTH])
+            player.attackRadius >
+              deltaEu - badDictionary[`baddie_${baddie.type}`].spriteWidth
           ) {
-            // if they are, the baddie is killed!
-            baddie.collisionStatus = 'hitFromFront';
-            console.log('kill successful at range', absDeltaX);
+            // if attack is successful, the baddie is killed!
+            baddie.collisionStatus = 'hit-from-front';
+            console.log(
+              'kill successful at relative distance ',
+              deltaEu - badDictionary[`baddie_${baddie.type}`].spriteWidth
+            );
+            // If your strike fails we must consider whether you can now be killed by them:
           } else if (
-            absDeltaX < badDictionary[baddie.id].faceToFace.killRadius
+            deltaEu <
+            badDictionary[`baddie_${baddie.type}`].faceToFace.killRadius
           ) {
-            player.collisionStatus = 'impact';
+            player.collisionStatus = `impact-with-${baddie.id}`;
+            console.log('face to face');
+            console.log(
+              `your position: ${player.x}, ${player.y}; their position: ${
+                baddie.x
+              }, ${baddie.y}, euclid: ${deltaEu}, attack rad: ${
+                player.attackRadius
+              }, thickness: ${baddie.spriteWidth}, relative distance: ${
+                deltaEu - baddie.spriteWidth
+              }`
+            );
           }
         } else {
-          console.log('you came from behind.');
+          // Player is to the right, player faces left, baddie faces left: player is behind baddie!
+          // This is the last case where the player stands to the right of a baddie and has the opportunity to attack:
           if (
             player.isAttacking &&
-            absDeltaX >
-              badDictionary[baddie.id].playerBehind.attackSuccessRange[0] &&
-            absDeltaX <
-              badDictionary[baddie.id].playerBehind.attackSuccessRange[1]
+            player.attackRadius >
+              deltaEu - badDictionary[`baddie_${baddie.type}`].spriteWidth
           ) {
-            // if they are, the baddie is killed!
-            baddie.collisionStatus = 'hitFromBehind';
-            console.log('kill successful at range', absDeltaX);
+            // if attack succeeds, the baddie is killed!
+            baddie.collisionStatus = 'hit-from-behind';
+            console.log(
+              'kill successful at relative distance ',
+              deltaEu - badDictionary[`baddie_${baddie.type}`].spriteWidth
+            );
           } else if (
-            absDeltaX < badDictionary[baddie.id].playerBehind.killRadius
+            deltaEu <
+            badDictionary[`baddie_${baddie.type}`].playerBehind.killRadius
           ) {
-            player.collisionStatus = 'impact';
+            player.collisionStatus = `impact-with-${baddie.id}`;
+            console.log('you came from behind.');
+            console.log(
+              `your position: ${player.x}, ${player.y}; their position: ${baddie.x}, ${baddie.y}, euclid: ${deltaEu}`
+            );
           }
         }
       }
     } else {
-      // If your delta X is less than zero you are left of the baddie:
+      // If your delta X is less than zero Player is LEFT of the baddie:
       if (player.facing === 'right') {
-        // If you are left of the baddie, facing right, and baddie faces right, you are behind him:
         if (baddie.facing === 'right') {
-          console.log('you came from behind.');
+          // Player is left of the baddie, facing right, and baddie faces right: player is behind baddie!
+          // Since you face the baddie you can attempt an attack:
           if (
             player.isAttacking &&
-            absDeltaX >
-              badDictionary[baddie.id].playerBehind.attackSuccessRange[0] &&
-            absDeltaX <
-              badDictionary[baddie.id].playerBehind.attackSuccessRange[1]
+            player.attackRadius >
+              deltaEu - badDictionary[`baddie_${baddie.type}`].spriteWidth
           ) {
             // if they are, the baddie is killed!
-            baddie.collisionStatus = 'hitFromBehind';
-            console.log('kill successful at range', absDeltaX);
+            baddie.collisionStatus = 'hit-from-behind';
+            console.log(
+              'kill successful at relative distance ',
+              deltaEu - badDictionary[`baddie_${baddie.type}`].spriteWidth
+            );
           } else if (
-            absDeltaX < badDictionary[baddie.id].playerBehind.killRadius
+            deltaEu <
+            badDictionary[`baddie_${baddie.type}`].playerBehind.killRadius
           ) {
-            player.collisionStatus = 'impact';
+            player.collisionStatus = `impact-with-${baddie.id}`;
+            console.log('you came from behind.');
+            console.log(
+              `your position: ${player.x}, ${player.y}; their position: ${baddie.x}, ${baddie.y}, euclid: ${deltaEu}`
+            );
           }
         } else {
-          console.log('face to face');
+          // Player is left of the baddie, player faces right, baddie faces left (head-on collision):
+          // Attempt attack:
           if (
             player.isAttacking &&
-            absDeltaX >
-              badDictionary[baddie.id].faceToFace.attackSuccessRange[0] &&
-            absDeltaX <
-              badDictionary[baddie.id].faceToFace.attackSuccessRange[1]
+            player.attackRadius >
+              deltaEu - badDictionary[`baddie_${baddie.type}`].spriteWidth
           ) {
             // if they are, the baddie is killed!
             baddie.collisionStatus = 'hitFromFront';
-            console.log('kill successful at range', absDeltaX);
+            console.log(
+              'kill successful at relative distance ',
+              deltaEu - badDictionary[`baddie_${baddie.type}`].spriteWidth
+            );
           } else if (
-            absDeltaX < badDictionary[baddie.id].faceToFace.killRadius
+            deltaEu <
+            badDictionary[`baddie_${baddie.type}`].faceToFace.killRadius
           ) {
-            player.collisionStatus = 'impact';
+            player.collisionStatus = `impact-with-${baddie.id}`;
+            console.log('face to face');
+            console.log(
+              `your position: ${player.x}, ${player.y}; their position: ${baddie.x}, ${baddie.y}, euclid: ${deltaEu}`
+            );
           }
         }
         // player faces left:
       } else {
+        // Player stands to the left, player faces left, baddie faces right: you face away from each other.
+        // No player attack is possible from this angle, so we go straight to seeing if they kill you:
         if (baddie.facing === 'right') {
-          console.log('back to back');
           if (
-            player.isAttacking &&
-            absDeltaX >
-              badDictionary[baddie.id].backToBack.attackSuccessRange[0] &&
-            absDeltaX <
-              badDictionary[baddie.id].backToBack.attackSuccessRange[1]
+            deltaEu <
+            badDictionary[`baddie_${baddie.type}`].backToBack.killRadius
           ) {
-            // if they are, the baddie is killed!
-            baddie.collisionStatus = 'hitFromFunnyAngle';
-            console.log('kill successful at range', absDeltaX);
-          } else if (
-            absDeltaX < badDictionary[baddie.id].backToBack.killRadius
-          ) {
-            player.collisionStatus = 'impact';
+            player.collisionStatus = `impact-with-${baddie.id}`;
+            console.log('back to back');
+            console.log(
+              `your position: ${player.x}, ${player.y}; their position: ${baddie.x}, ${baddie.y}, euclid: ${deltaEu}`
+            );
           }
         } else {
-          console.log('they came from behind!');
+          // Player stands to the left, player faces left, baddie faces left: baddie is behind you!
+          // No player attack possible:
           if (
-            player.isAttacking &&
-            absDeltaX >
-              badDictionary[baddie.id].baddieBehind.attackSuccessRange[0] &&
-            absDeltaX <
-              badDictionary[baddie.id].baddieBehind.attackSuccessRange[1]
+            deltaEu <
+            badDictionary[`baddie_${baddie.type}`].baddieBehind.killRadius
           ) {
-            // if they are, the baddie is killed!
-            baddie.collisionStatus = 'hitFromElsewhere';
-            console.log('kill successful at range', absDeltaX);
-          } else if (
-            absDeltaX < badDictionary[baddie.id].baddieBehind.killRadius
-          ) {
-            player.collisionStatus = 'impact';
+            player.collisionStatus = `impact-with-${baddie.id}`;
+            console.log('they came from behind!');
+            console.log(
+              `your position: ${player.x}, ${player.y}; their position: ${baddie.x}, ${baddie.y}, euclid: ${deltaEu}`
+            );
           }
         }
       }
@@ -176,71 +190,71 @@ class Collisions {
 
   // The compare method will assess the distance between the player and each baddie,
   // And call the DF method if the total distance (considering x and y) is within a given range:
-  compare(range) {
-    this.baddies.forEach((baddie) => {
-      if (!baddie.isDead) {
-        let xDist = this.player.x - baddie.x;
-        let yDist = this.player.y - baddie.y;
+  compare(range, baddies) {
+    baddies.forEach((baddie) => {
+      // the dead and dying can't hurt you (unless they come back as zombies, but that's in version 7):
+      if (!baddie.isDying && !baddie.isDead) {
+        const xDist = this.player.x - baddie.x;
+        const yDist = this.player.y - baddie.y;
+        const euclideanDist = (xDist ** 2 + yDist ** 2) ** 0.5;
         // When the player is within the proximity range, figure out the angles involved:
-        if ((xDist ** 2 + yDist ** 2) ** 0.5 < range) {
-          console.log((xDist ** 2 + yDist ** 2) ** 0.5);
-          // calculate the the (non-absolute) distance, and find which direction both sprites are facing:
+        if (euclideanDist < range) {
+          // let baddieRect = baddie.domElement.getBoundingClientRect();
+          // let goodieRect = this.player.domElement.getBoundingClientRect();
+          // console.log(
+          //   'your left edge: ',
+          //   goodieRect,
+          //   ' baddie right edge: ',
+          //   baddieRect
+          // );
+          // calculate the (signed) x-distance, and find which direction both sprites are facing:
           let deltaX = this.player.x - baddie.x;
-          this.determineFacing(deltaX, this.player, baddie);
+          this.determineFacing(deltaX, euclideanDist, this.player, baddie);
         }
       }
     });
   }
 }
 
-// The BAD reference matrix will tell you the outcomes to all manner of encounters with bad-guys.
-// Structure is as follows: Baddie type --> encounter angle --> {(attack?) + distance} --> outcome,
-// If the player is making an attack, check if the distance is within the attack's min/max range; if it isn't, or if the player
-// is not making an attack, check if the distance is within the kill radius. If the player is making an attack and the distance is
-// within the acceptable range, the baddie dies. If the player is outside of the successful attack range but within the kill radius,
-// the player dies. If neither range is applicable, then no outcome is produced.
+// The BAD reference matrix will tell you the outcomes to all angles of encounters with bad-guys.
+// The killRadius = the minimum Euclidean distance from a baddie that gets you killed.
+
 const badDictionary = {
   // For this baddie type,
   baddie_1001: {
+    // Sprite width is worth taking into account since some baddies will be wider than a single column:
+    spriteWidth: 1.125, // width is given in terms of columns
     // For this type of encounter,
     faceToFace: {
-      // attacks (if they're taking place and if they're allowed) must occur within a specified range to succeed:
-      attackSuccessRange: [0.5, 1.5],
-      // If no attack is made, or it is out of the 'goldilocks zone', use the kill radius to determine outcome:
-      killRadius: 1,
+      // use the kill radius to determine outcome:
+      killRadius: 0.9,
     },
     baddieBehind: {
-      attackSuccessRange: [0, 0], // an attack range of zero means that from this angle no attacks are possible.
-      killRadius: 1,
+      killRadius: 0.85,
     },
     playerBehind: {
-      attackSuccessRange: [0.5, 1.5], // Note that attacking this baddie from behind has a wider tolerance range for success.
-      killRadius: 1, // Note that attacking the stegosaurus from behind is riskier than it is with the foot soldier!
+      killRadius: 0.95, // Note that attacking the stegosaurus from behind is riskier than it is with the foot soldier!
     },
     backToBack: {
-      attackSuccessRange: [0, 0],
       killRadius: 1,
     },
   },
   baddie_1002: {
+    spriteWidth: 1,
     // For this type of encounter (initially there are four possible encounter angles),
     faceToFace: {
-      // attacks (if they're taking place and if they're allowed) must occur within a specified range to succeed:
-      attackSuccessRange: [0.5, 1.5],
-      // If no attack is made, or it is out of the 'goldilocks zone', use the kill radius to determine outcome:
+      // Use the kill radius to determine how close is too close for comfort:
       killRadius: 1,
     },
     baddieBehind: {
-      attackSuccessRange: [0, 0], // an attack range of zero means that from this angle no attacks are possible.
-      killRadius: 1,
+      killRadius: 0.8,
     },
     playerBehind: {
-      attackSuccessRange: [0.5, 1.5], // Note that attacking this baddie from behind has a wider tolerance range for success...
-      killRadius: 0.6, // ... and a narrower kill radius against you!
+      // Baddie kill radii change depend on what angle you're approaching from:
+      killRadius: 0.6,
     },
     bothFaceAway: {
-      attackSuccessRange: [0, 0],
-      killRadius: 0.6,
+      killRadius: 0.65,
     },
   },
 };